add outbox pattern

+++ catalog-service/src/main/java/com/ecommerce/catalogservice/dto/request/product/SkuItemForm.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -14,9 +15,9 @@ public class SkuItemForm {
     private String id;
     private String code;
     private String name;
-    private Double price;
-    private Double costPrice;
-    private Double originalPrice;
+    private BigDecimal price;
+    private BigDecimal costPrice;
+    private BigDecimal originalPrice;
     private Integer stock;
     private Boolean active;
     private MultipartFile image;
