+++ catalog-service/src/main/java/com/ecommerce/catalogservice/dto/response/product/ProductDetailDTO.java
@@ -8,6 +8,7 @@ import com.ecommerce.catalogservice.entity.ProductVariantGroup;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -23,13 +24,13 @@ public class ProductDetailDTO {
     private List<ProductVariantGroup> variantGroups;
     private List<ProductSpecDTO> specs;
     private Integer numOfReviews;
-    private Double avgRating;
+    private BigDecimal avgRating;
     private ImageEntity thumbnail;
     private List<ImageEntity> gallery;
     private String description;
     private String shortDescription;
     private ProductStatus status;
-    private Double minPrice;
-    private Double maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private List<SkuDetailDTO> skus;
 }
