+++ catalog-service/src/main/java/com/ecommerce/catalogservice/dto/response/product/SkuDTO.java
@@ -5,6 +5,7 @@ import com.ecommerce.catalogservice.entity.SkuSelect;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,8 +14,8 @@ public class SkuDTO {
     private String id;
     private String name;
     private String skuCode;
-    private Double price;
-    private Double originalPrice;
+    private BigDecimal price;
+    private BigDecimal originalPrice;
     private List<SkuSelect> selections;
     private ImageEntity thumbnail;
 }
